requirements.txt
asgiref==3.8.1
Django==5.2.3
django-admin-interface==0.30.1
django-adminlte3==0.1.6
django-appconf==1.1.0
django-colorfield==0.14.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-grappelli==4.0.2
django-jazzmin==3.0.1
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
python-slugify==8.0.4
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
